Fixed enlightened bug

+++ src/main/java/lefishe/main/effects/EnlightenedStatusEffect.java
@@ -18,7 +18,7 @@ public class EnlightenedStatusEffect extends StatusEffect {
 	@Override
 	public boolean canApplyUpdateEffect(int duration, int amplifier) {
 		//if effect is about to end, delete all light blocks
-		if (duration < 3){
+		if (duration < 3 && lightBlocks != null){
 			lightBlocks.deleteAll();
 			return false;
 		}
